requirements.txt
openai
torch
PyPDF2
pyyaml
beautifulsoup4
lxml
python-dotenv
numpy